+++ component/src/test/java/arez/component/ContainerTest.java
@@ -369,7 +369,7 @@ public class ContainerTest
     }
 
     @Nonnull
-    protected String getContainerName()
+    protected String getName()
     {
       return "MyRepository";
     }