+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   whether an entity is detached or disposed when the container is disposed. 
 * **\[component\]** Add invariant check to `AbstractContainer.attach(entity)` to ensure that entity
   is not already attached.
+* **\[component\]** Rename `AbstractContainer.getComponentName()` to `AbstractContainer.getName()`.
+* **\[component\]** Extract `AbstractEntryContainer` from `AbstractContainer` to simplify creating
+  other utilities that need to have references to entities removed when the entity is disposed.
 
 ### [v0.89](https://github.com/arez/arez/tree/v0.89) (2018-06-07)
 [Full Changelog](https://github.com/arez/arez/compare/v0.88...v0.89)